feat(rollup): Create vehicle in Advance function
This commit adds the functionality to create a vehicle in the `Advance` function of the `router.go` file. The new case "create_vehicle" is implemented, calling the `CreateVehicle` function from the `advance` package. This change allows the creation of a new vehicle record in the database when the corresponding action is triggered.

Note: This commit message follows the established convention of prefixing the commit type (`feat`) followed by the affected module (`rollup`) in parentheses.